Time: 64 ms (6.93%), Space: 90.7 MB (6.81%) - LeetHub

+++ 0004-median-of-two-sorted-arrays/0004-median-of-two-sorted-arrays.cpp
@@ -15,7 +15,7 @@ public:
                 return v[n/2];
             }
             else{
-                return double((v[n/2] + v[n/2-1]))/2;
+                return ((v[n/2] + v[n/2-1]))/2;
             
         }
